requirements.txt
Brotli==1.0.7
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
dash==1.13.3
dash-bootstrap-components==0.10.2
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.0
dash-table==4.8.1
DateTime==4.3
decorator==4.4.2
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
networkx==2.4
numpy==1.18.5
pandas==1.0.4
plotly==4.8.1
python-dateutil==2.8.1
python-igraph==0.8.2
pytz==2020.1
requests==2.23.0
retrying==1.3.3
scipy==1.4.1
six==1.15.0
texttable==1.6.2
urllib3==1.25.9
Werkzeug==1.0.1
zope.interface==5.1.0
